requirements.txt
numpy
mss
Pillow
pytesseract
PyQt5
QtWebEngine
screeninfo
opencv-python-headless